drop worker namespace
