+++ spatialos-sdk/src/snapshot.rs
@@ -1,4 +1,4 @@
-use crate::worker::{entity::Entity, utils::cstr_to_string, EntityId};
+use crate::{entity::Entity, utils::cstr_to_string, EntityId};
 use spatialos_sdk_sys::worker::*;
 use std::{ffi::CString, path::Path};
